# src/main/java/com/xm/technical/cryptoinvestment/model/CryptoDetails.java
package com.xm.technical.cryptoinvestment.model;

import lombok.AllArgsConstructor;
import lombok.Data;

import java.math.BigDecimal;

@Data
@AllArgsConstructor
public class CryptoDetails {
    private String cryptoName;
    private BigDecimal oldest;
    private BigDecimal newest;
    private BigDecimal min;
    private BigDecimal max;
}